backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
psycopg==3.1.8
watchdog==5.0.3
PyMuPDF==1.24.12
pandas==2.2.3
numpy==2.0.2
pymupdf4llm
xlsxwriter
pytesseract
pillow
pymupdf
pdfplumber
pdf2image
llama-index


